Explicitely use uint256

+++ contracts/AccessControl/RoleBasedAccessControl.sol
@@ -161,7 +161,7 @@ contract RoleBasedAccessControl is
         require(subjects.length == roles.length);
         require(subjects.length == objects.length);
         require(subjects.length == newValues.length);
-        for(uint i = 0; i < subjects.length; i++) {
+        for(uint256 i = 0; i < subjects.length; i++) {
             setUserRolePrivate(subjects[i], roles[i], objects[i], newValues[i]);
         }
     }
@@ -222,7 +222,7 @@ contract RoleBasedAccessControl is
             list.push(subject);
         }
         if(oldValue != TriState.Unset && newValue == TriState.Unset) {
-            for(uint i = 0; i < list.length; i++) {
+            for(uint256 i = 0; i < list.length; i++) {
                 if(list[i] == subject) {
                     list[i] = list[list.length - 1];
                     delete list[list.length - 1];
